fix: split deepseek_v3_fp4 ut
