+++ test/srt/run_suite.py
@@ -155,6 +155,7 @@ suites = {
     "per-commit-4-gpu-b200": [
         # TestFile("test_gpt_oss_4gpu.py", 600),
         TestFile("test_deepseek_v3_fp4_4gpu.py", 3600),
+        TestFile("test/srt/test_deepseek_v3_fp4_cutlass_4gpu.py", 1200),
     ],
     "per-commit-4-gpu-deepep": [
         TestFile("ep/test_deepep_small.py", 531),
